+++ lib/ironman/config.ex
@@ -46,7 +46,7 @@ defmodule Ironman.Config do
             changed: MapSet.new(),
             skipped_upgrades: MapSet.new()
 
-  def new!() do
+  def new! do
     %__MODULE__{
       mix_exs: file_or_nil(Utils.path_of(:mix_exs)),
       config_exs: file_or_nil(Utils.path_of(:config_exs)),